+++ ui/appui-layout-react/.eslintrc.js
@@ -1,6 +1,0 @@
-require("@rushstack/eslint-patch/modern-module-resolution");
-
-module.exports = {
-  extends: ["../.eslintrc.base.js", "prettier"],
-  parserOptions: { tsconfigRootDir: __dirname },
-};