+++ ui/appui-layout-react/src/appui-layout-react/state/ToolSettingsState.ts
@@ -6,16 +6,26 @@
  * @module Base
  */
 
-/** @internal */
-export const toolSettingsTabId = "nz-tool-settings-tab";
+import produce from "immer";
+import { UiError } from "@itwin/appui-abstract";
+import { category } from "./internal/NineZoneStateHelpers";
+import type { NineZoneState } from "./NineZoneState";
+import type { TabState } from "./TabState";
+import { removeTabFromWidget } from "./TabState";
+import { getTabLocation } from "./TabLocation";
 
 /** @internal */
-export interface DockedToolSettingsState {
+export interface CommonToolSettingsState {
+  readonly tabId: TabState["id"];
+}
+
+/** @internal */
+export interface DockedToolSettingsState extends CommonToolSettingsState {
   readonly type: "docked";
 }
 
 /** @internal */
-export interface WidgetToolSettingsState {
+export interface WidgetToolSettingsState extends CommonToolSettingsState {
   readonly type: "widget";
 }
 
@@ -24,9 +34,74 @@ export type ToolSettingsState =
   | DockedToolSettingsState
   | WidgetToolSettingsState;
 
-/** @internal */
-export function isDockedToolSettingsState(
-  state: ToolSettingsState
-): state is DockedToolSettingsState {
-  return state.type === "docked";
+/** Adds a docked tool settings.
+ * @internal
+ */
+export function addDockedToolSettings(
+  state: NineZoneState,
+  tabId: TabState["id"]
+): NineZoneState {
+  if (state.toolSettings)
+    throw new UiError(category, "Tool settings already exist");
+  if (!(tabId in state.tabs))
+    throw new UiError(category, "Tab does not exist", undefined, () => ({
+      tabId,
+    }));
+  const location = getTabLocation(state, tabId);
+  if (location)
+    throw new UiError(
+      category,
+      "Tab is already in a widget",
+      undefined,
+      () => ({ tabId, widgetId: location.widgetId })
+    );
+
+  return produce(state, (stateDraft) => {
+    stateDraft.toolSettings = {
+      tabId,
+      type: "docked",
+    };
+  });
+}
+
+/** Adds a widget to the tool settings.
+ * @internal
+ */
+export function addWidgetToolSettings(
+  state: NineZoneState,
+  tabId: TabState["id"]
+): NineZoneState {
+  if (state.toolSettings)
+    throw new UiError(category, "Tool settings already exist");
+  if (!(tabId in state.tabs))
+    throw new UiError(category, "Tab does not exist", undefined, () => ({
+      tabId,
+    }));
+  const location = getTabLocation(state, tabId);
+  if (!location)
+    throw new UiError(category, "Tab is not in a widget", undefined, () => ({
+      tabId,
+    }));
+
+  return produce(state, (stateDraft) => {
+    stateDraft.toolSettings = {
+      tabId,
+      type: "widget",
+    };
+  });
+}
+
+/** Removes tab from the tool settings, but keeps the tab state.
+ * @internal
+ */
+export function removeToolSettings(state: NineZoneState): NineZoneState {
+  if (!state.toolSettings) return state;
+
+  if (state.toolSettings.type === "widget") {
+    state = removeTabFromWidget(state, state.toolSettings.tabId);
+  }
+
+  return produce(state, (draft) => {
+    draft.toolSettings = undefined;
+  });
 }