+++ ui/appui-react/src/appui-react/widget-panels/Frontstage.tsx
@@ -31,10 +31,12 @@ import type {
   TabState,
 } from "@itwin/appui-layout-react";
 import {
+  addDockedToolSettings,
   addFloatingWidget,
   addPanelWidget,
   addTab,
   addTabToWidget,
+  addWidgetToolSettings,
   convertAllPopupWidgetContainersToFloating,
   createLayoutStore,
   createNineZoneState,
@@ -54,7 +56,7 @@ import {
   panelSides,
   removeTab,
   removeTabFromWidget,
-  toolSettingsTabId,
+  removeToolSettings,
   WidgetPanels,
 } from "@itwin/appui-layout-react";
 import type {
@@ -69,7 +71,6 @@ import type {
   TabLocation,
   WidgetDef,
   WidgetEventArgs,
-  WidgetStateChangedEventArgs,
 } from "../widgets/WidgetDef";
 import { WidgetContent } from "./Content";
 import { WidgetPanelsFrontstageContent } from "./FrontstageContent";
@@ -220,20 +221,7 @@ export function useNineZoneDispatch(frontstageDef: FrontstageDef) {
       if (action.type === "RESIZE") {
         InternalFrontstageManager.nineZoneSize = Size.create(action.size);
       }
-      // istanbul ignore if
-      if (action.type === "TOOL_SETTINGS_DRAG_START") {
-        void UiFramework.postTelemetry(
-          "Tool Settings Undocking",
-          "28B04E07-AE73-4533-A0BA-8E2A8DC99ADF"
-        );
-      }
-      // istanbul ignore if
-      if (action.type === "TOOL_SETTINGS_DOCK") {
-        void UiFramework.postTelemetry(
-          "Tool Settings Docking to Settings Bar",
-          "BEDE684B-B3DB-4637-B3AF-DC3CBA223F94"
-        );
-      }
+
       if (action.type === "WIDGET_TAB_POPOUT") {
         const tabId = action.id;
         frontstageDef.popoutWidget(tabId);
@@ -373,6 +361,22 @@ export function useLabels() {
   );
 }
 
+function toTabArgs(widget: WidgetDef) {
+  const label = getWidgetLabel(widget.label);
+  return {
+    allowedPanelTargets: widget.allowedPanelTargets?.map((location) =>
+      toPanelSide(location)
+    ),
+    hideWithUiWhenFloating: !!widget.hideWithUiWhenFloating,
+    canPopout: widget.canPopout,
+    iconSpec: widget.iconSpec,
+    isFloatingStateWindowResizable: widget.isFloatingStateWindowResizable,
+    label,
+    preferredFloatingWidgetSize: widget.defaultFloatingSize,
+    preferredPanelWidgetSize: widget.preferredPanelSize,
+  };
+}
+
 /** @internal */
 export function addWidgets(
   state: NineZoneState,
@@ -387,19 +391,7 @@ export function addWidgets(
 
   const tabs = new Array<string>();
   for (const widget of visibleWidgets) {
-    const label = getWidgetLabel(widget.label);
-    state = addTab(state, widget.id, {
-      label,
-      iconSpec: widget.iconSpec,
-      preferredPanelWidgetSize: widget.preferredPanelSize,
-      preferredFloatingWidgetSize: widget.defaultFloatingSize,
-      canPopout: widget.canPopout,
-      isFloatingStateWindowResizable: widget.isFloatingStateWindowResizable,
-      hideWithUiWhenFloating: !!widget.hideWithUiWhenFloating,
-      allowedPanelTargets: widget.allowedPanelTargets?.map((location) =>
-        toPanelSide(location)
-      ),
-    });
+    state = addTab(state, widget.id, toTabArgs(widget));
     tabs.push(widget.id);
   }
 
@@ -424,13 +416,10 @@ export function appendWidgets(
   location: StagePanelLocation,
   section: StagePanelSection
 ): NineZoneState {
-  if (widgetDefs.length === 0) return state;
-
   const side = toPanelSide(location);
 
   // Add missing widget tabs.
   for (const widgetDef of widgetDefs) {
-    const label = getWidgetLabel(widgetDef.label);
     const saveTab = state.tabs[widgetDef.id];
     const preferredPanelWidgetSize = saveTab
       ? saveTab.preferredPanelWidgetSize
@@ -445,17 +434,10 @@ export function appendWidgets(
           !!saveTab.preferredFloatingWidgetSize)
       : !!widgetDef.defaultFloatingSize;
     state = addTab(state, widgetDef.id, {
-      label,
-      iconSpec: widgetDef.iconSpec,
-      canPopout: widgetDef.canPopout,
-      preferredPanelWidgetSize,
+      ...toTabArgs(widgetDef),
       preferredFloatingWidgetSize,
+      preferredPanelWidgetSize,
       userSized,
-      isFloatingStateWindowResizable: widgetDef.isFloatingStateWindowResizable,
-      hideWithUiWhenFloating: !!widgetDef.hideWithUiWhenFloating,
-      allowedPanelTargets: widgetDef.allowedPanelTargets?.map((l) =>
-        toPanelSide(l)
-      ),
     });
 
     if (
@@ -653,6 +635,20 @@ export function addMissingWidgets(
     StagePanelSection.End
   );
 
+  if (frontstageDef.toolSettings) {
+    const newToolSettingsWidgets = determineNewWidgets(
+      [frontstageDef.toolSettings],
+      state
+    );
+    for (const newToolSettingsWidget of newToolSettingsWidgets) {
+      state = addTab(
+        state,
+        newToolSettingsWidget.id,
+        toTabArgs(newToolSettingsWidget)
+      );
+    }
+  }
+
   return state;
 }
 
@@ -679,9 +675,9 @@ function hideWidgets(
  */
 export function removeMissingWidgets(
   frontstageDef: FrontstageDef,
-  initialState: NineZoneState
+  state: NineZoneState
 ): NineZoneState {
-  let state = initialState;
+  const toolSettingsTabId = state.toolSettings?.tabId;
   for (const [, tab] of Object.entries(state.tabs)) {
     if (tab.id === toolSettingsTabId) continue;
     const widgetDef = frontstageDef.findWidgetDef(tab.id);
@@ -842,7 +838,8 @@ function getPanelMaxSize(
   return (maxSizeSpec.percentage / 100) * size;
 }
 
-const stateVersion = 13; // this needs to be bumped when NineZoneState is changed (to recreate the layout).
+/** @internal */
+export const stateVersion = 14; // this needs to be bumped when NineZoneState is changed (to recreate the layout).
 
 /** @internal */
 export function initializeNineZoneState(
@@ -866,41 +863,40 @@ export function initializeNineZoneState(
     frontstageDef,
     StagePanelLocation.Bottom
   );
-  nineZone = produce(nineZone, (stateDraft) => {
-    for (const [, panel] of Object.entries(stateDraft.panels)) {
+  nineZone = produce(nineZone, (draft) => {
+    for (const [, panel] of Object.entries(draft.panels)) {
       const expanded = panel.widgets.find(
-        (widgetId) => stateDraft.widgets[widgetId].minimized === false
+        (widgetId) => draft.widgets[widgetId].minimized === false
       );
       const firstWidget =
-        panel.widgets.length > 0
-          ? stateDraft.widgets[panel.widgets[0]]
-          : undefined;
+        panel.widgets.length > 0 ? draft.widgets[panel.widgets[0]] : undefined;
       // istanbul ignore next
       if (!expanded && firstWidget) {
         firstWidget.minimized = false;
       }
     }
-    stateDraft.panels.left.collapsed = isPanelCollapsed(
+    draft.panels.left.collapsed = isPanelCollapsed(
       frontstageDef.leftPanel?.panelState
     );
-    stateDraft.panels.right.collapsed = isPanelCollapsed(
+    draft.panels.right.collapsed = isPanelCollapsed(
       frontstageDef.rightPanel?.panelState
     );
-    stateDraft.panels.top.collapsed = isPanelCollapsed(
+    draft.panels.top.collapsed = isPanelCollapsed(
       frontstageDef.topPanel?.panelState
     );
-    stateDraft.panels.bottom.collapsed = isPanelCollapsed(
+    draft.panels.bottom.collapsed = isPanelCollapsed(
       frontstageDef.bottomPanel?.panelState
     );
-
-    const toolSettingsWidgetDef = frontstageDef.toolSettings;
-    if (toolSettingsWidgetDef) {
-      const toolSettingsTab = stateDraft.tabs[toolSettingsTabId];
-      toolSettingsTab.preferredPanelWidgetSize =
-        toolSettingsWidgetDef.preferredPanelSize;
-    }
   });
+
   nineZone = addMissingWidgets(frontstageDef, nineZone);
+
+  const toolSettingsWidgetDef = frontstageDef.toolSettings;
+  if (toolSettingsWidgetDef) {
+    const toolSettingsTabId = toolSettingsWidgetDef.id;
+    nineZone = addDockedToolSettings(nineZone, toolSettingsTabId);
+  }
+
   nineZone = hideWidgets(nineZone, frontstageDef);
   nineZone = processPopoutWidgets(nineZone, frontstageDef);
 
@@ -918,7 +914,7 @@ export function restoreNineZoneState(
 ): NineZoneState {
   let restored: NineZoneState = {
     ...saved,
-    tabs: defaultNineZone.tabs,
+    tabs: {},
   };
 
   for (const [, tab] of Object.entries(saved.tabs)) {
@@ -941,15 +937,9 @@ export function restoreNineZoneState(
     restored = produce(restored, (draft) => {
       draft.tabs[tab.id] = {
         ...tab,
-        label: getWidgetLabel(widgetDef.label),
-        hideWithUiWhenFloating: widgetDef.hideWithUiWhenFloating,
-        iconSpec: widgetDef.iconSpec,
-        canPopout: widgetDef.canPopout,
-        isFloatingStateWindowResizable:
-          widgetDef.isFloatingStateWindowResizable,
-        allowedPanelTargets: widgetDef.allowedPanelTargets?.map((location) =>
-          toPanelSide(location)
-        ),
+        ...toTabArgs(widgetDef),
+        preferredFloatingWidgetSize: tab.preferredFloatingWidgetSize,
+        preferredPanelWidgetSize: tab.preferredPanelWidgetSize,
       };
     });
   }
@@ -1026,12 +1016,9 @@ export function packNineZoneState(state: NineZoneState): SavedNineZoneState {
   };
   packed = produce(packed, (draft) => {
     for (const [, tab] of Object.entries(state.tabs)) {
-      if (tab.id === toolSettingsTabId) continue;
-
       draft.tabs[tab.id] = {
         id: tab.id,
         preferredFloatingWidgetSize: tab.preferredFloatingWidgetSize,
-        allowedPanelTargets: tab.allowedPanelTargets,
         userSized: tab.userSized,
       };
     }
@@ -1203,10 +1190,21 @@ export function setWidgetState(
     return hideWidget(state, widgetDef);
   } else if (widgetState === WidgetState.Open) {
     const id = widgetDef.id;
+    if (widgetDef.isToolSettings) {
+      if (!state.toolSettings && !widgetDef.tabLocation) {
+        state = addDockedToolSettings(state, id);
+      }
+      if (state.toolSettings?.type === "docked") return state;
+    }
+
     let location = getTabLocation(state, id);
     if (!location) {
       state = addHiddenWidget(state, widgetDef);
       location = getTabLocation(state, id);
+
+      if (widgetDef.isToolSettings) {
+        state = addWidgetToolSettings(state, id);
+      }
     }
 
     return produce(state, (draft) => {
@@ -1231,6 +1229,10 @@ export function setWidgetState(
     });
   } else if (widgetState === WidgetState.Closed) {
     const id = widgetDef.id;
+    if (widgetDef.isToolSettings && state.toolSettings?.type !== "widget") {
+      return state;
+    }
+
     let location = getTabLocation(state, id);
     if (!location) {
       state = addHiddenWidget(state, widgetDef);
@@ -1263,14 +1265,35 @@ export function setWidgetState(
     state = removeTabFromWidget(state, id);
     const bounds = widgetDef.tabLocation?.floatingWidget?.bounds;
     state = addFloatingWidget(state, getUniqueId(), [id], { bounds });
+
+    if (widgetDef.isToolSettings) {
+      if (state.toolSettings?.type === "docked") {
+        state = removeToolSettings(state);
+      }
+      if (!state.toolSettings) {
+        state = addWidgetToolSettings(state, id);
+      }
+    }
   }
   return state;
 }
 
 /** Stores widget location and hides it in the UI. */
 function hideWidget(state: NineZoneState, widgetDef: WidgetDef) {
+  const toolSettingsTabId = state.toolSettings?.tabId;
+
+  if (widgetDef.id === toolSettingsTabId) {
+    widgetDef.tabLocation = undefined;
+    state = produce(state, (draft) => {
+      draft.toolSettings = undefined;
+    });
+
+    if (state.toolSettings?.type === "docked") return state;
+  }
+
   const location = getTabLocation(state, widgetDef.id);
   if (!location) return state;
+
   // istanbul ignore else
   if (isFloatingTabLocation(location)) {
     const floatingWidget = state.floatingWidgets.byId[location.widgetId];
@@ -1386,8 +1409,8 @@ export function useSavedFrontstageState(frontstageDef: FrontstageDef) {
       if (isFrontstageStateSettingResult(settingResult)) {
         const setting = settingResult.setting;
         if (
-          setting.version >= version &&
-          setting.stateVersion >= stateVersion
+          setting.version === version &&
+          setting.stateVersion === stateVersion
         ) {
           const restored = restoreNineZoneState(
             frontstageDef,
@@ -1508,10 +1531,9 @@ export function useFrontstageManager(
   useToolAsToolSettingsLabel?: boolean
 ) {
   React.useEffect(() => {
-    const listener = createListener(
-      frontstageDef,
-      ({ widgetDef, widgetState }: WidgetStateChangedEventArgs) => {
-        assert(!!frontstageDef.nineZoneState);
+    return UiFramework.frontstages.onWidgetStateChangedEvent.addListener(
+      ({ widgetDef, widgetState }) => {
+        if (!frontstageDef.nineZoneState) return;
         frontstageDef.nineZoneState = setWidgetState(
           frontstageDef.nineZoneState,
           widgetDef,
@@ -1519,12 +1541,6 @@ export function useFrontstageManager(
         );
       }
     );
-    UiFramework.frontstages.onWidgetStateChangedEvent.addListener(listener);
-    return () => {
-      UiFramework.frontstages.onWidgetStateChangedEvent.removeListener(
-        listener
-      );
-    };
   }, [frontstageDef]);
   React.useEffect(() => {
     const listener = createListener(
@@ -1620,6 +1636,8 @@ export function useFrontstageManager(
       const label = useToolAsToolSettingsLabel
         ? IModelApp.tools.find(toolId)?.flyover || defaultLabel
         : defaultLabel;
+      const toolSettingsTabId = frontstageDef.nineZoneState.toolSettings?.tabId;
+      if (!toolSettingsTabId) return;
       frontstageDef.nineZoneState = setWidgetLabel(
         frontstageDef.nineZoneState,
         toolSettingsTabId,