+++ ui/core-react/src/core-react/messagebox/MessageBox.tsx
@@ -11,7 +11,6 @@ import classnames from "classnames";
 import * as React from "react";
 import type { DialogButtonDef } from "@itwin/appui-abstract";
 import { MessageSeverity } from "@itwin/appui-abstract";
-import { SimpleDialog } from "../dialog/SimpleDialog";
 import type { CommonProps } from "../utils/Props";
 import {
   SvgHelpCircular,
@@ -28,6 +27,7 @@ import {
 } from "@itwin/itwinui-icons-react";
 import type { IconSpec } from "../icons/IconComponent";
 import { Icon } from "../icons/IconComponent";
+import { Dialog } from "../dialog/Dialog";
 
 /** Properties for the [[MessageBox]] component
  * @public
@@ -82,7 +82,7 @@ export class MessageBox extends React.PureComponent<MessageBoxProps> {
 
   public override render(): React.ReactElement {
     return (
-      <SimpleDialog
+      <Dialog
         title={this.props.title}
         buttonCluster={this.props.buttonCluster}
         opened={this.props.opened}
@@ -100,7 +100,7 @@ export class MessageBox extends React.PureComponent<MessageBoxProps> {
         >
           {this.props.children}
         </MessageContainer>
-      </SimpleDialog>
+      </Dialog>
     );
   }
 }