+++ ui/appui-react/CHANGELOG.md
@@ -1,6 +1,19 @@
 # Change Log - @itwin/appui-react
 
-This log was last generated on Tue, 27 Jun 2023 14:50:22 GMT and should not be manually modified.
+This log was last generated on Mon, 17 Jul 2023 10:16:20 GMT and should not be manually modified.
+
+## 4.3.0
+Mon, 17 Jul 2023 10:16:20 GMT
+
+### Updates
+
+- Update sinon to 15.2.0 and @types/sinon to 10.0.15
+- Reword emtpy tool settings message to include tool name.
+- Use UiFramework.store instead of StateManager.store to display popout widgets.
+- Fix empty tool settings message.
+- Fix `FrontstageDef.findWidgetDef` to correctly return `toolSettings` and `statusBar` widgets.
+- Fix `WidgetDef.setWidgetState` to correctly hide and show tool settings.
+- Add missing `isMounted` check to fix a warning message while running tests.
 
 ## 4.2.0
 Tue, 27 Jun 2023 14:50:22 GMT